# openapi/api.yaml
openapi: 3.0.2
servers:
- description: Ecom Test API Service (GAE)
  url: https://open247-gae.appspot.com
- description: Ecom Test API Service
  url: http://localhost:8080
- description: Ecom Test API Service Ngrok
  url: https://a7c46957.ngrok.io
info:
  description: |
    # Ecom API
    Most requests require a JSON Web Token (JWT) aquired by signing into the Firebase Auth that corresponds to the API endpoint.

    ## Authentication
    API resources are protected by levels of access:

      * None (No JWT required)
      * `RoleShopper` (JWT contains no custom `role` claim)
      * `RoleCustomer` (JWT contains a custom claim `"role": "customer"`)
      * `RoleAdmin` (JWT contains a custom claim `"role": "admin"`)
      * `RoleSuperUser` (JWT contains a custom claim `"role": "root"`)

    Set the `Authorization: Bearer $JWT` header before calling the API endpoints that require authentication.

  version: "v0.57.0-dev-1809"
  title: Headless Ecommerce API
  contact:
    name: Andy Fusniak
    email: andy@andyfusniak.com
paths:
...
  /carts:
    post:
      security:
      - bearerAuth: []
      summary: Create a new shopping cart
      description: |
        OpCreateCart requires `RoleShopper` privileges. Creates an empty shopping cart. A shopping cart consists of items, offers and coupons. The request body contains a unique identifier for the new cart.
      operationId: OpCreateCart
      tags:
      - Carts
      responses:
        '201':
          description: cart object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
              example:
                object: cart
                id: '30ad2997-3d19-4001-88d9-e2568d8cf720'
                locked: false
                created: '2019-08-02T12:02:42.217936Z'
                modified: 2019-08-02T12:02:42.217936Z'
                items:
                  object: 'list'
                  data: []
...
  /carts/{id}/items:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identifier for the cart.
      schema:
        type: string
        format: uuid
        example: '30ad2997-3d19-4001-88d9-e2568d8cf720'
    get:
      security:
      - bearerAuth: []
      summary: Get cart items
      description: |
        OpGetCartItems requires `RoleShopper` privileges.
      operationId: OpGetCartItems
      tags:
      - Cart Items
      responses:
        '200':
          description: List of cart item objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: 'list'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
        '404':
          description: |
            The cart with cart_id in the path could not be found. You should first create a cart calling the `OpCreateCart` operation and use its cart_id.
          content:
            application/json:
              example: ''
    delete:
      security:
      - bearerAuth: []
      summary: Empty the entire cart
      description: |
        OpEmptyCartItems requires `RoleShopper` privileges.
      operationId: OpEmptyCartItems
      tags:
      - Cart Items
      responses:
        '204':
          description: No Content
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cart/cart-contains-no-items:
                  summary: cart/cart-contains-no-items
                  value:
                    status: 409
                    code: 'cart/cart-contains-no-items'
                    message: OpEmptyCartItems cannot be called if a cart contains no items
  /carts-items:
    post:
      security:
      - bearerAuth: []
      summary: Place an item into a cart
      description: |
        OpAddItemToCart requires `RoleShopper` privileges. Places an new item to the cart returning the cart_item object. Each item has a unique identifier.
      operationId: OpAddItemToCart
      tags:
      - Cart Items
      requestBody:
        description: New CartItem
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - cart_id
              - product_id
              - qty
              properties:
                cart_id:
                  type: string
                  format: uuid
                  example: 'd9b210a3-0a0a-485f-8851-0ca77cee02f8'
                product_id:
                  type: string
                  format: uuid
                  example: '54bfd164-b31a-4e61-b715-9c1cbe3c18e3'
                qty:
                  type: integer
                  example: 2
      responses:
        '201':
          description: |
            `CartItem` object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
        '404':
          description: |
            The cart with cart_id in the path could not be found. You should first create a cart calling the `OpCreateCart` operation and use its cart_id.
          content:
            application/json:
              example: ''
        '409':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cart/cart-item-already-exists:
                  summary: cart/cart-item-already-exists
                  value:
                    status: 409
                    code: 'cart/cart-item-already-exists'
                    message: cart item already exists in the cart
  /carts-items/{id}:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identifier for the cart item.
      schema:
        type: string
        example: 'aba8b505-d016-4afc-95bd-1f292ed5a915'
    patch:
      security:
      - bearerAuth: []
      summary: Update a cart item with a new quantity
      description: |
        OpUpdateCartItem requires `RoleShopper` privileges.
      operationId: OpUpdateCartItem
      tags:
      - Cart Items
      requestBody:
        description: New quantity for cart item
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - qty
              properties:
                qty:
                  type: integer
                  example: 3
      responses:
        '200':
          description: CartItem object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
    delete:
      security:
      - bearerAuth: []
      summary: Delete a cart item
      description: |
        OpDeleteCartItem requires `RoleShopper` privileges.
      operationId: OpDeleteCartItem
      tags:
      - Cart Items
      responses:
        '204':
          description: No Content
...
  /promo-rules:
    post:
      security:
      - bearerAuth: []
      summary: Create a new promotion rule
      description: |
        OpCreatePromoRule requires `RoleAdmin` privileges.
      operationId: OpCreatePromoRule
      tags:
      - Promotion Rules
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoRuleCreateRequest'
      responses:
        '201':
          description: promo_rule object
          content:
            applicatio/json:
              schema:
                $ref: '#/components/schemas/PromoRule'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      security:
      - bearerAuth: []
      summary: Get a list of all promotion rules
      description: |
        OpListPromoRules requires `RoleAdmin` privileges.
      operationId: OpListPromoRules
      tags:
      - Promotion Rules
      responses:
        '200':
          description: list of promo_list objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: 'list'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PromoRule'
  /promo-rules/{id}:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identifier for the promo rule.
      schema:
        type: string
        format: uuid
        example: 'd8f3d5d4-9143-4ec0-87fe-0753afecaa87'
    get:
      security:
      - bearerAuth: []
      summary: Get a promotion rules
      description: |
        OpGetPromoRule requires `RoleAdmin` privileges.
      operationId: OpGetPromoRule
      tags:
      - Promotion Rules
      responses:
        '200':
          description: promo_rule object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromoRule'
    delete:
      security:
      - bearerAuth: []
      summary: Delete a promotion rules
      description: |
        OpDeletePromoRule requires `RoleAdmin` privileges.
      operationId: OpDeletePromoRule
      tags:
      - Promotion Rules
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
...
  /customers:
    post:
      summary: Create and sign up a new customer
      description: |
        Sign up a new customer creates a new customer with the role of `RoleCustomer`.
        ### Auth
        OpCreateCustomer requires no JSON Web Token.
      operationId: OpCreateCustomer
      tags:
      - Customers
      requestBody:
        description: New customer details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    get:
      security:
      - bearerAuth: []
      summary: List all customers
      description: |
        OpListCustomers requires `RoleAdmin` privileges.
      operationId: OpListCustomers
      tags:
      - Customers
      responses:
        '200':
          description: List of Customer objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
  /categories:
    put:
      security:
      - bearerAuth: []
      summary: Replace all categories
      description: |
        OpUpdateCatalog requires `RoleAdmin` privileges.
      operationId: OpUpdateCatalog
      tags:
      - Categories
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
            example:
              segment: a
              name: ประเภท A
              categories:
              - segment: b
                name: ประเภท B
                categories:
                - segment: e
                  name: ประเภท E
                  categories: []
              - segment: c
                name: ประเภท C
                categories:
                - segment: f
                  name: ประเภท F
                  categories:
                  - segment: i
                    name: ประเภท I
                    categories: []
                  - segment: j
                    name: ประเภท J
                    categories:
                    - segment: m
                      name: ประเภท M
                      categories: []
                    - segment: 'n'
                      name: ประเภท N
                      categories: []
                - segment: g
                  name: ประเภท G
                  categories: []
              - segment: d
                name: ประเภท D
                categories:
                - segment: h
                  name: ประเภท H
                  categories:
                  - segment: k
                    name: ประเภท K
                    categories: []
                  - segment: l
                    name: ประเภท L
                    categories: []
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cart/cart-item-already-exists:
                  summary: cart/cart-item-already-exists
                  value:
                    status: 409
                    code: 'cart/cart-item-already-exists'
                    message: cart item already exists in the cart
    get:
      security:
      - bearerAuth: []
      summary: Get all categories
      description: |
        OpGetCatalog requires `RoleShopper` privileges.
      operationId: OpGetCatalog
      tags:
      - Categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categories'
    delete:
      security:
      - bearerAuth: []
      summary: Purge all categories
      description: |
        OpPurgeCatalog requires `RoleAdmin` privileges.
      operationId: OpPurgeCatalog
      tags:
      - Categories
      responses:
        '204':
          description: No Content
...
  /customers/{id}:
    get:
      security:
      - bearerAuth: []
      summary: Get a customer object
      description: |
        OpGetCustomer requires `RoleCustomer` privileges.
      operationId: OpGetCustomer
      tags:
      - Customers
      parameters:
      - name: id
        required: true
        in: path
        description: Customer ID
        schema:
          type: string
          example: '53b3384b-fc35-47d1-9054-6b81f30f382e'
      responses:
        '201':
          description: Customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{id}/devkeys:
    parameters:
    - name: id
      required: true
      in: path
      description: Customer ID
      schema:
        type: string
        example: '53b3384b-fc35-47d1-9054-6b81f30f382e'
    post:
      security:
      - bearerAuth: []
      summary: Generate a new developer key
      description: |
        OpGenerateCustomerDevKey requires `RoleCustomer` privileges.
      operationId: OpGenerateCustomerDevKey
      tags:
      - Developer Keys
      responses:
        '201':
          description: Developer Key object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevKey'
    get:
      security:
      - bearerAuth: []
      summary: List a customer's developer keys
      description: |
        OpListCustomersDevKeys requires `RoleCustomer` privileges.
      operationId: OpListCustomersDevKeys
      tags:
      - Developer Keys
      responses:
        '200':
          description: List of Developer Key objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DevKey'
  /devkeys/{id}:
    delete:
      security:
      - bearerAuth: []
      summary: Delete a developer key
      description: |
        OpDeleteCustomerDevKey requires `RoleAdmin` privileges.
      operationId: OpDeleteCustomerDevKey
      tags:
      - Developer Keys
      parameters:
      - name: id
        required: true
        in: path
        description: Developer Key ID
        schema:
          type: string
          example: '5a85466e-9f23-4d71-b8b7-adbfc479bb6c'
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  /signin-with-devkey:
    post:
      summary: Signin with a developer key
      description: |
        OpSignInWithDevKey requires no JWT Web Token.

        Signin with a secret developer key. After you receive the custom token in the response, pass it to Firebase Auth [signInWithCustomToken](https://firebase.google.com/docs/reference/node/firebase.auth.Auth#sign-inwith-custom-token) to sign in the user:


        ```javascript
          firebase.auth().signInWithCustomToken(token).catch(function(error) {
            // Handle Errors here.
            var errorCode = error.code;
            var errorMessage = error.message;
            // ...
          });
        ```

        To sign out a user, call `signOut`:

        ```javascript
          firebase.auth().signOut().then(function() {
            // Sign-out successful.
          }).catch(function(error) {
            // An error happened.
          });
        ```
      operationId: OpSignInWithDevKey
      tags:
      - Developer Keys
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInDevKeyRequest'
      responses:
        '201':
          description: Custom Token object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInDevKeyCustomToken'
  /customers/{id}/addresses:
    parameters:
    - name: id
      required: true
      in: path
      description: Customer ID
      schema:
        type: string
        example: '7835fa6e-38de-4959-8838-c88ab3e073b3'
    post:
      security:
      - bearerAuth: []
      summary: Add a new address for a given customer
      description: |
        OpCreateAddress requires `RoleCustomer` privileges.
      operationId: OpCreateAddress
      tags:
      - Addresses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '201':
          description: Address object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    get:
      security:
      - bearerAuth: []
      summary: Get all addresses for a given customer
      description: |
        OpGetCustomersAddresses requires `RoleCustomer` privileges. Retrieves a list of Address objects for the given customer. If called with `RoleCustomer` you must be the customer you intend to retrieve the addresses for. If called with `RoleAdmin` you can retrieve a list of addresses for any customer.
      operationId: OpGetCustomersAddresses
      tags:
      - Addresses
      responses:
        '200':
          description: List of Address objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
  /addresses/{id}:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identfier for the address.
      schema:
        type: string
        example: 'e3afa4f4-c5de-4e0f-8c37-5be677cfde6a'
    patch:
      security:
      - bearerAuth: []
      summary: Update an address
      description: |
        OpUpdateAddress requires `RoleCustomer` privileges.
      operationId: OpUpdateAddress
      tags:
      - Addresses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: Address object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    get:
      security:
      - bearerAuth: []
      summary: Get an address object
      description: |
        OpGetAddress requires `RoleCustomer` privileges.
      operationId: OpGetAddress
      tags:
      - Addresses
      responses:
        '200':
          description: Address object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      security:
      - bearerAuth: []
      summary: Delete an address object
      description: |
        OpDeleteAddress requires `RoleCustomer` privileges.
      operationId: OpDeleteAddress
      tags:
      - Addresses
      responses:
        '204':
          description: No Content
...
  /products/{id}/prices:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identifier for the product.
      schema:
        type: string
        format: uuid
        example: 'd9da65ce-6f2c-42ae-8789-448f3053b185'
    put:
      parameters:
      - name: price_list_id
        required: true
        in: query
        description: filter for only a single price list
        schema:
          type: string
          format: uuid
          example: '72ae053a-e63d-4a8e-8a0b-ac43e9ca89d1'
      security:
      - bearerAuth: []
      summary: Get all prices for a given product
      description: |
        OpUpdateProductPrices requires `RoleShopper` privileges.
      operationId: OpUpdateProductPrices
      tags:
      - Prices
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductPriceRequest'
      responses:
        '200':
          description: list of price objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: 'list'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Price'
    get:
      parameters:
      - name: price_list_id
        required: false
        in: query
        description: filter for only a single price list
        schema:
          type: string
          format: uuid
          example: '72ae053a-e63d-4a8e-8a0b-ac43e9ca89d1'
      security:
      - bearerAuth: []
      summary: Get all prices for a given product
      description: |
        OpGetProductPrices requires `RoleShopper` privileges.
      operationId: OpGetProductPrices
      tags:
      - Prices
      responses:
        '200':
          description: list of price objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: 'list'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Price'
    delete:
      security:
      - bearerAuth: []
      summary: Delete a product's prices by price list id.
      description: |
        OpDeleteTierPricing requires `RoleAdmin` privileges.
      operationId: OpDeleteTierPricing
      tags:
      - Prices
      responses:
        '204':
          description: No Content
  /products/{id}/images:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identifier for the product.
      schema:
        type: string
        format: uuid
        example: 'd9da65ce-6f2c-42ae-8789-448f3053b185'
    post:
      security:
      - bearerAuth: []
      summary: Add an image to a product
      description: |
        OpAddImage requires `RoleAdmin` privileges.
      operationId: OpAddImage
      tags:
      - Images
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageRequest'
      responses:
        '201':
          description: Image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
    get:
      security:
      - bearerAuth: []
      summary: List product images
      description: |
        OpListProductImages requires `RoleShopper` privileges.
      operationId: OpListProductImages
      tags:
      - Images
      responses:
        '200':
          description: List of image objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: 'list'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Image'
    delete:
      security:
      - bearerAuth: []
      summary: Delete all product images for a given product SKU
      description: |
        OpDeleteAllProductImages requires `RoleAdmin` privileges.
      operationId: OpDeleteAllProductImages
      tags:
      - Images
      responses:
        '204':
          description: No Content
  /images/{id}:
    parameters:
    - name: id
      required: true
      in: path
      description: A unique identifier for the image.
      schema:
        type: string
        example: '390d6786-0ff7-43e7-844b-af816589a045'
    get:
      security:
      - bearerAuth: []
      summary: Get an image by image ID
      description: |
        OpGetImage requires `RoleShopper` privileges.
      operationId: OpGetImage
      tags:
      - Images
      responses:
        '200':
          description: Image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
    delete:
      security:
      - bearerAuth: []
      summary: Delete an image by Image ID
      description: |
        OpDeleteImage requires `RoleAdmin` privileges.
      operationId: OpDeleteImage
      tags:
      - Images
      responses:
        '204':
          description: No Content
...
  /products:
    post:
      security:
      - bearerAuth: []
      summary: Create a new product
      description: |
        OpCreateProduct requires `RoleAdmin` privileges.
      operationId: OpCreateProduct
      tags:
      - Products
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '409':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                product/product-path-exists:
                  summary: product/product-path-exists
                  value:
                    status: 409
                    code: 'product/product-path-exists'
                    message: product path already exists
    get:
      security:
      - bearerAuth: []
      summary: List all products
      description: |
        OpListProducts requires `RoleShopper` privileges.
      operationId: OpListProducts
      tags:
      - Products
      responses:
        '200':
          description: Product object
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: 'list'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
...
  /sysinfo:
    get:
      security:
      - bearerAuth: []
      summary: Get system information
      description: |
        OpSystemInfo requires `RoleAdmin` privileges.
      operationId: OpSystemInfo
      tags:
      - Misc
      responses:
        '200':
          description: System info object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfo'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Cart:
      type: object
      properties:
        object:
          type: string
          example: 'cart'
        id:
          type: string
          example: '30ad2997-3d19-4001-88d9-e2568d8cf720'
        locked:
          type: string
          example: false
        created:
          type: string
          format: date-time
          example: '2019-08-02T12:02:42.217936Z'
        modified:
          type: string
          format: date-time
          example: '2019-08-02T12:02:42.217936Z'
        items:
          type: object
          properties:
            object:
              type: string
              example: 'list'
            data:
              type: array
              items:
                $ref: '#/components/schemas/CartItem'
...